Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main() {
     int a,b,c;
     float avg;
     scanf("%d %d %d",&a,&b,&c);
-    int x=a+b+c;
+    float x=a+b+c;
     printf("%d",x);
     avg = (x)/3;
     printf("Average: %.2f", avg);
